backend/requirements.txt
Django>=4.2,<4.3
psycopg2-binary
djangorestframework
mongoengine
django-cors-headers
python-dotenv
